Create NavigationQueueContainer
- create NavigationQueueContainer
- extract queue logic from navigationService and add it to queue contianer
