+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/NavigationQueueContainer.java
@@ -1,0 +1,176 @@
+package com.mapbox.services.android.navigation.v5.navigation;
+
+import android.location.Location;
+
+import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
+import com.mapbox.services.android.navigation.v5.utils.RingBuffer;
+import com.mapbox.services.android.navigation.v5.utils.time.TimeUtils;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.Iterator;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
+
+class NavigationQueueContainer {
+  private static final int TWENTY_SECOND_INTERVAL = 20;
+  private static final String MOCK_PROVIDER = "com.mapbox.services.android.navigation.v5.location.MockLocationEngine";
+
+  private List<SessionState> queuedRerouteEvents = new ArrayList<>();
+  private List<FeedbackEvent> queuedFeedbackEvents = new ArrayList<>();
+  private MapboxNavigation mapboxNavigation;
+  String locationEngineName;
+  private RouteProgress routeProgress;
+  private Location currentLocation;
+  private RingBuffer<Location> locationBuffer;
+
+  NavigationQueueContainer(MapboxNavigation mapboxNavigation) {
+    this.mapboxNavigation = mapboxNavigation;
+    locationBuffer = new RingBuffer<>(20);
+  }
+
+  void setCurrentLocation(Location rawLocation) {
+    this.currentLocation = rawLocation;
+    locationBuffer.addLast(rawLocation);
+  }
+
+  void sendQueues() {
+    for (FeedbackEvent feedbackEvent : queuedFeedbackEvents) {
+      sendFeedbackEvent(feedbackEvent);
+    }
+    for (SessionState sessionState : queuedRerouteEvents) {
+      sendRerouteEvent(sessionState);
+    }
+  }
+
+  void rerouteSessionsStateUpdate() {
+    mapboxNavigation.getEventDispatcher().onUserOffRoute(currentLocation);
+    mapboxNavigation.setSessionState(mapboxNavigation.getSessionState().toBuilder()
+      .eventLocation(currentLocation).build());
+  }
+
+  public void rerouteOccurred() {
+    mapboxNavigation.setSessionState(mapboxNavigation.getSessionState().toBuilder()
+      .routeProgressBeforeReroute(routeProgress)
+      .beforeRerouteLocations(Arrays.asList(
+        locationBuffer.toArray(new Location[locationBuffer.size()])))
+      .previousRouteDistancesCompleted(
+        mapboxNavigation.getSessionState().previousRouteDistancesCompleted()
+          + routeProgress.distanceTraveled())
+      .rerouteDate(new Date())
+      .build());
+    queuedRerouteEvents.add(mapboxNavigation.getSessionState());
+  }
+
+  String recordFeedbackEvent(String feedbackType, String description,
+                                  @FeedbackEvent.FeedbackSource String feedbackSource) {
+    SessionState feedbackEventSessionState = mapboxNavigation.getSessionState().toBuilder()
+      .rerouteDate(new Date())
+      .beforeRerouteLocations(Arrays.asList(
+        locationBuffer.toArray(new Location[locationBuffer.size()])))
+      .routeProgressBeforeReroute(routeProgress)
+      .eventLocation(currentLocation)
+      .mockLocation((currentLocation.getProvider().equals(MOCK_PROVIDER)) ? true : false)
+      .build();
+
+    FeedbackEvent feedbackEvent = new FeedbackEvent(feedbackEventSessionState, feedbackSource);
+    feedbackEvent.setDescription(description);
+    feedbackEvent.setFeedbackType(feedbackType);
+    queuedFeedbackEvents.add(feedbackEvent);
+
+    return feedbackEvent.getFeedbackId();
+  }
+
+  void updateFeedbackEvent(String feedbackId,
+                                  @FeedbackEvent.FeedbackType String feedbackType, String description) {
+    FeedbackEvent feedbackEvent = findQueuedFeedbackEvent(feedbackId);
+    if (feedbackEvent != null) {
+      feedbackEvent.setFeedbackType(feedbackType);
+      feedbackEvent.setDescription(description);
+    }
+  }
+
+  void cancelFeedback(String feedbackId) {
+    FeedbackEvent feedbackEvent = findQueuedFeedbackEvent(feedbackId);
+    queuedFeedbackEvents.remove(feedbackEvent);
+  }
+
+  void sendFeedbackEvent(FeedbackEvent feedbackEvent) {
+    SessionState feedbackSessionState = feedbackEvent.getSessionState();
+    feedbackSessionState = feedbackSessionState.toBuilder().afterRerouteLocations(Arrays.asList(
+      locationBuffer.toArray(new Location[locationBuffer.size()])))
+      .build();
+
+    NavigationMetricsWrapper.feedbackEvent(feedbackSessionState, routeProgress,
+      feedbackEvent.getSessionState().eventLocation(), feedbackEvent.getDescription(),
+      feedbackEvent.getFeedbackType(), "", feedbackEvent.getFeedbackId(),
+      mapboxNavigation.obtainVendorId(), locationEngineName);
+  }
+
+  void sendRerouteEvent(SessionState sessionState) {
+    sessionState = sessionState.toBuilder()
+      .afterRerouteLocations(Arrays.asList(
+        locationBuffer.toArray(new Location[locationBuffer.size()])))
+      .build();
+
+    NavigationMetricsWrapper.rerouteEvent(sessionState, routeProgress,
+      sessionState.eventLocation(), locationEngineName);
+
+    for (SessionState session : queuedRerouteEvents) {
+      queuedRerouteEvents.set(queuedRerouteEvents.indexOf(session),
+        session.toBuilder().lastRerouteDate(
+          sessionState.rerouteDate()
+        ).build());
+    }
+
+    mapboxNavigation.setSessionState(mapboxNavigation.getSessionState().toBuilder().lastRerouteDate(
+      sessionState.rerouteDate()
+    ).build());
+  }
+
+  private FeedbackEvent findQueuedFeedbackEvent(String feedbackId) {
+    for (FeedbackEvent feedbackEvent : queuedFeedbackEvents) {
+      if (feedbackEvent.getFeedbackId().equals(feedbackId)) {
+        return feedbackEvent;
+      }
+    }
+    return null;
+  }
+
+  void updateFeedbackQueue() {
+    Iterator<FeedbackEvent> iterator = queuedFeedbackEvents.listIterator();
+    while (iterator.hasNext()) {
+      FeedbackEvent feedbackEvent = iterator.next();
+      if (feedbackEvent.getSessionState().eventLocation() != null
+        && feedbackEvent.getSessionState().eventLocation().equals(locationBuffer.peekFirst())
+        || TimeUtils.dateDiff(feedbackEvent.getSessionState().rerouteDate(), new Date(), TimeUnit.SECONDS)
+        > TWENTY_SECOND_INTERVAL) {
+        sendFeedbackEvent(feedbackEvent);
+        iterator.remove();
+      }
+    }
+  }
+
+  void updateRerouteQueue() {
+    Iterator<SessionState> iterator = queuedRerouteEvents.listIterator();
+    while (iterator.hasNext()) {
+      SessionState sessionState = iterator.next();
+      if (sessionState.eventLocation() != null
+        && sessionState.eventLocation().equals(locationBuffer.peekFirst())
+        || TimeUtils.dateDiff(sessionState.rerouteDate(), new Date(), TimeUnit.SECONDS)
+        > TWENTY_SECOND_INTERVAL) {
+        sendRerouteEvent(sessionState);
+        iterator.remove();
+      }
+    }
+  }
+
+  void setLocationEngineName(String locationEngineName) {
+    this.locationEngineName = locationEngineName;
+  }
+
+  void setRouteProgress(RouteProgress routeProgress) {
+    this.routeProgress = routeProgress;
+  }
+}
